# lib/api-zod/src/generated/types/createWeightCheckBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sthenos Fitness App API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateWeightCheckBody {
  weightKg: number;
  date: string;
  /** @nullable */
  notes?: string | null;
}