# lib/api-zod/src/generated/types/dailyReport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sthenos Fitness App API
 * OpenAPI spec version: 0.1.0
 */

export interface DailyReport {
  date: string;
  totalCaloriesBurned: number;
  totalCaloriesConsumed: number;
  netCalories: number;
  exerciseCount: number;
  mealCount: number;
  breakfastCalories: number;
  lunchCalories: number;
  dinnerCalories: number;
  snackCalories: number;
}