# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sthenos Fitness App API
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  id: number;
  name: string;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  weightKg?: number | null;
  /** @nullable */
  heightCm?: number | null;
  /** @nullable */
  goal?: string | null;
  /** @nullable */
  activityLevel?: string | null;
  /** @nullable */
  bmi?: number | null;
  /** @nullable */
  bmr?: number | null;
  /** @nullable */
  tdee?: number | null;
  createdAt: string;
  updatedAt: string;
}